+++ app/api/payMerchant/route.js
@@ -1,6 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
-import Order from '../../models/Order';
-import { isAdmin } from '../../utils/functions';
+import Order from '../../../models/Order';
+import { isAdmin } from '../../../utils/functions';
 import { NextResponse } from 'next/server';
 
 const CHAPA_SECRET_KEY = "CHASECK_TEST-s6oBbGS04bRkcXLT7P6x2do2EKcCXfJ6";
