+++ models/SuperAdmin.js
@@ -1,0 +1,32 @@
+import mongoose from "mongoose";
+
+const superAdminSchema = new mongoose.Schema(
+  {
+    email: {
+      type: String,
+      required: true,
+      unique: true,
+    },
+    fullname: {
+      type: String,
+      required: true,
+    },
+    password: {
+      type: String,
+      required: true,
+    },
+    phone: {
+      type: String,
+      default: null,
+    },
+    role: {
+      type: String,
+      default: "superAdmin"
+    },
+  },
+  { timestamps: true } 
+);
+
+const SuperAdmin = mongoose.models.SuperAdmin || mongoose.model("SuperAdmin", superAdminSchema);
+
+export default SuperAdmin;
